--- a/src/ImageSharp/Dithering/ErrorDiffusion/ErrorDiffuser.cs
+++ b/src/ImageSharp/Dithering/ErrorDiffusion/ErrorDiffuser.cs
@@ -5,12 +5,11 @@
 
 namespace ImageSharp.Dithering
 {
-    using System;
     using System.Numerics;
     using System.Runtime.CompilerServices;
 
     /// <summary>
-    /// The base class for performing effor diffusion based dithering.
+    /// The base class for performing error diffusion based dithering.
     /// </summary>
     public abstract class ErrorDiffuser : IErrorDiffuser
     {
@@ -35,18 +34,23 @@ namespace ImageSharp.Dithering
         private readonly int startingOffset;
 
         /// <summary>
+        /// The diffusion matrix
+        /// </summary>
+        private readonly Fast2DArray<float> matrix;
+
+        /// <summary>
         /// Initializes a new instance of the <see cref="ErrorDiffuser"/> class.
         /// </summary>
         /// <param name="matrix">The dithering matrix.</param>
         /// <param name="divisor">The divisor.</param>
-        protected ErrorDiffuser(Fast2DArray<float> matrix, byte divisor)
+        internal ErrorDiffuser(Fast2DArray<float> matrix, byte divisor)
         {
             Guard.NotNull(matrix, nameof(matrix));
             Guard.MustBeGreaterThan(divisor, 0, nameof(divisor));
 
-            this.Matrix = matrix;
-            this.matrixWidth = this.Matrix.Width;
-            this.matrixHeight = this.Matrix.Height;
+            this.matrix = matrix;
+            this.matrixWidth = this.matrix.Width;
+            this.matrixHeight = this.matrix.Height;
             this.divisorVector = new Vector4(divisor);
 
             this.startingOffset = 0;
@@ -61,9 +65,6 @@ namespace ImageSharp.Dithering
                 }
             }
         }
-
-        /// <inheritdoc />
-        public Fast2DArray<float> Matrix { get; }
 
         /// <inheritdoc />
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -87,9 +88,9 @@ namespace ImageSharp.Dithering
 
                     if (matrixX > 0 && matrixX < width && matrixY > 0 && matrixY < height)
                     {
-                        float coefficient = this.Matrix[row, col];
+                        float coefficient = this.matrix[row, col];
 
-                        // Good to disable here as we are not comparing matematical output.
+                        // Good to disable here as we are not comparing mathematical output.
                         // ReSharper disable once CompareOfFloatsByEqualityOperator
                         if (coefficient == 0)
                         {