--- a/src/ImageSharp/Processing/Processors/ColorMatrix/GrayscaleBt601Processor.cs
+++ b/src/ImageSharp/Processing/Processors/ColorMatrix/GrayscaleBt601Processor.cs
@@ -9,11 +9,11 @@ namespace ImageSharp.Processing.Processors
     using System.Numerics;
 
     /// <summary>
-    /// Converts the colors of the image to Grayscale applying the formula as specified by
-    /// ITU-R Recommendation BT.601 <see href="https://en.wikipedia.org/wiki/Luma_%28video%29#Rec._601_luma_versus_Rec._709_luma_coefficients"/>.
+    /// Converts the colors of the image to Grayscale applying the formula as specified by ITU-R Recommendation BT.601
+    /// <see href="https://en.wikipedia.org/wiki/Luma_%28video%29#Rec._601_luma_versus_Rec._709_luma_coefficients"/>.
     /// </summary>
     /// <typeparam name="TColor">The pixel format.</typeparam>
-    public class GrayscaleBt601Processor<TColor> : ColorMatrixFilter<TColor>
+    internal class GrayscaleBt601Processor<TColor> : ColorMatrixProcessor<TColor>
         where TColor : struct, IPixel<TColor>
     {
         /// <inheritdoc/>