--- a/src/ImageSharp/Processing/Processors/ColorMatrix/GrayscaleBt709Processor.cs
+++ b/src/ImageSharp/Processing/Processors/ColorMatrix/GrayscaleBt709Processor.cs
@@ -9,11 +9,11 @@ namespace ImageSharp.Processing.Processors
     using System.Numerics;
 
     /// <summary>
-    /// Converts the colors of the image to Grayscale applying the formula as specified by
-    /// ITU-R Recommendation BT.709 <see href="https://en.wikipedia.org/wiki/Rec._709#Luma_coefficients"/>.
+    /// Converts the colors of the image to Grayscale applying the formula as specified by ITU-R Recommendation BT.709
+    /// <see href="https://en.wikipedia.org/wiki/Rec._709#Luma_coefficients"/>.
     /// </summary>
     /// <typeparam name="TColor">The pixel format.</typeparam>
-    public class GrayscaleBt709Processor<TColor> : ColorMatrixFilter<TColor>
+    internal class GrayscaleBt709Processor<TColor> : ColorMatrixProcessor<TColor>
         where TColor : struct, IPixel<TColor>
     {
         /// <inheritdoc/>