--- a/src/ImageSharp/Quantizers/IQuantizer.cs
+++ b/src/ImageSharp/Quantizers/IQuantizer.cs
@@ -5,8 +5,6 @@
 
 namespace ImageSharp.Quantizers
 {
-    using System;
-
     using ImageSharp.Dithering;
 
     /// <summary>