… README.md
… src/ImageSharp/Processing/Processors/ColorMatrix/SaturationProcessor.cs
… src/ImageSharp/Processing/Processors/ColorMatrix/SepiaProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/BoxBlurProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/Convolution2DProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/Convolution2PassProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/ConvolutionProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/EdgeDetector2DProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/EdgeDetectorCompassProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/EdgeDetectorProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/KayyaliProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/KirschProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/Laplacian3X3Processor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/Laplacian5X5Processor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/LaplacianOfGaussianProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/PrewittProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/RobertsCrossProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/RobinsonProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/ScharrProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/EdgeDetection/SobelProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/GaussianBlurProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/GaussianSharpenProcessor.cs
… src/ImageSharp/Processing/Processors/Effects/AlphaProcessor.cs
… src/ImageSharp/Processing/Processors/Effects/BackgroundColorProcessor.cs
… src/ImageSharp/Processing/Processors/Effects/BrightnessProcessor.cs
… src/ImageSharp/Processing/Processors/Effects/ContrastProcessor.cs
… src/ImageSharp/Processing/Processors/Effects/InvertProcessor.cs
… src/ImageSharp/Processing/Processors/Effects/OilPaintingProcessor.cs
… src/ImageSharp/Processing/Processors/Effects/PixelateProcessor.cs
… src/ImageSharp/Processing/Processors/Overlays/GlowProcessor.cs