### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/JimBobSquarePants/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practise as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
This PR internalizes all the `IImageProcessor<TColor>` implementations as well as `PixelArea<TColor>` and `PixelDataPool<T>`, and `Fast2DArray<T>`.

<strike>I would have liked to internalize `Fast2DArray<T>` but that would mean changing `IErrorDiffusor`. </strike>

<strike>Answers on a postcard! </strike>

<!-- Thanks for contributing to ImageSharp! -->