@antonfirsov 2017-03-17
Looks good to me! 
PixelArea is internal, all memory utils are internal, processors are internal ... can't think of anything else.