@antonfirsov 2017-03-17
OK found one another thing: 
I think [protected virtual methods in Quantizer base class](https://github.com/JimBobSquarePants/ImageSharp/blob/master/src/ImageSharp/Quantizers/Octree/Quantizer.cs#L166). Per-pixel virtual calls lead to instant cache misses, I don't think users should hook into them.