@JimBobSquarePants 2017-03-17
@antonfirsov I managed to refactor that method. Here's some benchmarks. 

**Before**

Method | LargeImage |        Mean |    StdErr |     StdDev |      Median | Scaled | Scaled-StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
---------------------------------- |----------- |------------ |---------- |----------- |------------ |------- |-------------- |--------- |--------- |--------- |---------- |
'ImageSharp Octree Png' |      False | 353.8687 ms | 3.1196 ms | 12.0823 ms | 352.1816 ms |   1.00 |          0.00 | 800.0000 | 583.3333 | 237.5000 |   9.16 MB |
'ImageSharp Octree NoDither Png' |      False |  76.2996 ms | 0.7592 ms |  6.7054 ms |  78.5256 ms |   0.22 |          0.02 | 412.7907 | 225.2907 |        - |   3.08 MB |
'ImageSharp Palette Png' |      False | 434.3553 ms | 4.4461 ms | 37.4638 ms | 423.4140 ms |   1.23 |          0.11 | 425.5952 | 288.6905 | 232.1429 |   6.85 MB |
'ImageSharp Palette NoDither Png' |      False |  86.8364 ms | 1.1569 ms | 11.5107 ms |  86.4912 ms |   0.25 |          0.03 | 417.5000 | 292.5000 | 230.0000 |   1.99 MB |
'ImageSharp Wu Png' |      False | 144.5337 ms | 1.4385 ms | 13.8720 ms | 142.1108 ms |   0.41 |          0.04 |        - |        - |        - | 872.71 kB |


**After**

Method | LargeImage |        Mean |    StdDev | Scaled | Scaled-StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
---------------------------------- |----------- |------------ |---------- |------- |-------------- |--------- |--------- |--------- |---------- |
'ImageSharp Octree Png' |      False | 234.0621 ms | 2.1119 ms |   1.00 |          0.00 | 708.3333 | 491.6667 | 145.8333 |   9.16 MB |
'ImageSharp Octree NoDither Png' |      False |  44.4091 ms | 0.1494 ms |   0.19 |          0.00 |  73.5294 |        - |        - |   3.08 MB |
'ImageSharp Palette Png' |      False | 289.4772 ms | 2.8749 ms |   1.24 |          0.02 | 354.9107 | 209.8214 | 158.4821 |   6.85 MB |
'ImageSharp Palette NoDither Png' |      False |  56.1580 ms | 0.1950 ms |   0.24 |          0.00 |        - |        - |        - |   1.99 MB |
'ImageSharp Wu Png' |      False |  97.9099 ms | 1.1120 ms |   0.42 |          0.01 |        - |        - |        - | 872.75 kB |